Added more ways to add/delete things
- Way to add a performance from the performances page
Done. I removed the link from the artists pages since it doesn't make
sense there with the way we have performances set up now.

- Way to add a venue from the venue page
Done.

- Way to delete venue
Done. It's on the venue details page now. I noticed if you delete a
venue that has a performance attached to it and look at the performance
page, the setlist disappears.